Changes to models - Moved code that dealt with files to folder fileData'
