+++ controllers/adminController.js
@@ -1,4 +1,4 @@
-const Product = require('../models/product');
+const Product = require('../models/fileData/product');
 
 getProducts = (req, res, next) => {
